==> 20_Preprocessing.ipynb <==
# 결측치 대체 / 삭제

- 명목형 변수들은 결측치를 'NaN' 이라는 문자열으로 대체 

- 연속형 변수들은 주변 칼럼을 참조하여 최대한 추정, 해결 안된 결측치들은 평균 대체

- int나 float값을 가지지만 명목 변수인 칼럼들도 존재, 이러한 칼럼들 수정

- 이후 과정은 module을 사용하여 진행 (Preprocessing 모듈의 Preprocess 클래스 참조)

def syspath(path) :
    if path not in sys.path :
        sys.path.append(path)
    else : pass

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import sys
import warnings

warnings.filterwarnings('ignore')

syspath('./dataset/mod')

from Preprocessing import Preprocess

# 중복만 없앤 원본 데이터 불러오기
df_origin = pd.read_pickle('./dataset/00_feature_with_label.pkl')
df_origin.head()

# 전체 null값 개수와 그 중 LABEL 1인 것들의 개수를 출력하는 함수 생성
def get_result(column_name) :
    mask1 = df_origin['{}'.format(column_name)].isnull()
    tmp_df = df_origin[mask1]
    mask2 = tmp_df.LABEL == 1
    
    print('##### {} #####'.format(column_name))
    print('Number of null : {}'.format(len(tmp_df)))
    print('Label 1 null : {}'.format(len(tmp_df[mask2])))
    return tmp_df

# 연속형 변수들을 함수에 대입하여 전체 null값 개수와 그 중 LABEL이 1인 것들의 개수를 확인

columns =  ['CUS_TOTAL_NUMBER_OF_PACKAGES_6', 'VAL_FINANCIAL_VALUE_12',
            'TOT_FINANCIAL_VALUE_22', 'IMP_EXCHANGE_RATE_23',
            'IMP_INLAND_TRANSPORT_MODE_25', 'IMP_TRANSPORT_MODE_AT_BODR_26',
            'GDS_GROSS_MASS_35', 'IDG_NET_MASS_38', 'FIN_FINANCIAL_VALUE_42',
            'STC_FINANCIAL_VALUE_46']

for column in columns : 
    get_result(column)

# 결과를 살펴보면 label = 1인 null값은 한개도 없음
# 따라서 null값의 수가 많은 column들(500개 이상)만 대체 고려,

## IMP_EXCHANGE_RATE_23
- TOT_CURRENCY_22 = TJS 인 sample은 null값을 1로 대체 (자국 화폐이므로)

mask = (df_origin.TOT_CURRENCY_22 == 'TJS') & (df_origin.IMP_EXCHANGE_RATE_23.isnull())
idx = df_origin[mask].index
for i in idx :
    df_origin.iloc[i, 32] = 1
df_origin[mask]['IMP_EXCHANGE_RATE_23'].unique()

## IMP_INLAND_TRANSPORT_MODE_25, IMP_TRANSPORT_MODE_AT_BODR_26

# 출력값에서 볼 수 있듯이 이 둘은 float값을 지니고 있지만 사실상 명목 변수
print(df_origin.IMP_INLAND_TRANSPORT_MODE_25.unique())
print(df_origin.IMP_TRANSPORT_MODE_AT_BODR_26.unique())

# null값 0으로 대체 후 int로 변환
for column in ['IMP_INLAND_TRANSPORT_MODE_25', 'IMP_TRANSPORT_MODE_AT_BODR_26'] :
    df_origin['{}'.format(column)].fillna(0, inplace=True)
    df_origin['{}'.format(column)] = df_origin['{}'.format(column)].astype(int)
    print(df_origin['{}'.format(column)].unique())

## GDS_GROSS_MASS_35, IDG_NET_MASS_38

# 해당하는 칼럼들만 추출하여 DataFrame 생성
edf = df_origin[['COM_COMBINED_NOMENCLATURE_33','GDS_GROSS_MASS_35', 'IDG_NET_MASS_38']]
edf

# null값 대체를 위한 사용자 함수 생성

def translation(column_name) :

  # 해당 칼럼의 null값이 있는 샘플만 인덱싱하여 해당 샘플의 인덱스와 HS-CODE를 리스트로 저장 
  mask = edf[column_name].isnull()
  idx_list = edf[mask].index.tolist()
  hs_list = edf[mask]['COM_COMBINED_NOMENCLATURE_33'].unique().tolist()

  # HS-CODE별로 GDS_GROSS_MASS_35, IDG_NET_MASS_38의 평균 산출 후 dictionary 생성 
  mean_list = []
  for code in hs_list :
    mask = edf['COM_COMBINED_NOMENCLATURE_33'] == code
    mean = edf[mask][column_name].mean()
    mean_list.append(mean)
  mean_dict = dict(zip(hs_list, mean_list))

  # dictionary를 사용하여 null값 대체 
  for i in idx_list :
    val = edf.iloc[i, 0]
    if column_name == 'GDS_GROSS_MASS_35' :
      edf.iloc[i, 1] = mean_dict[val]
    elif column_name == 'IDG_NET_MASS_38' :
      edf.iloc[i, 2] = mean_dict[val]

for column in ['GDS_GROSS_MASS_35', 'IDG_NET_MASS_38'] :
  translation(column)

for column in ['GDS_GROSS_MASS_35', 'IDG_NET_MASS_38'] :
  df_origin[column] = edf[column]

# null값이 100개 대로 줄어든 것을 확인 가능
columns =  ['CUS_TOTAL_NUMBER_OF_PACKAGES_6', 'VAL_FINANCIAL_VALUE_12',
            'TOT_FINANCIAL_VALUE_22', 'IMP_EXCHANGE_RATE_23',
            'IMP_INLAND_TRANSPORT_MODE_25', 'IMP_TRANSPORT_MODE_AT_BODR_26',
            'GDS_GROSS_MASS_35', 'IDG_NET_MASS_38', 'FIN_FINANCIAL_VALUE_42',
            'STC_FINANCIAL_VALUE_46']

for column in columns : 
    get_result(column)

# 나머지 null값은 이전과 같이 평균 대체 진행
pre_df_origin = Preprocess(df_origin)
basic_feature_df = pre_df_origin.null_solution()
basic_feature_df.isnull().sum()

# 결측값이 모두 제거된 것을 확인 가능

# 결측값 제거된 DataFrame을 pickle 파일로 저장
basic_feature_df.to_pickle('./dataset/20_basic_feature_df.pkl')
