==> 21_VIF.ipynb <==
# VIF 산출

- 변수간 다중 공선성을 감안, 이를 해결하기 위한 지표를 산출

- 모든 변수의 VIF가 20 이하가 될 때까지 VIF를 반복산출

- GDS_GROSS_MASS_35 / COR_FINANCIAL_VALUE / IMP_TRANSPORT_MODE_AT_BODR_26 / TRD_COUNTRY_2 4개 칼럼 삭제

import numpy as np
import pandas as pd
import warnings
import sys

warnings.filterwarnings('ignore')

def syspath(path) :
    if path not in sys.path :
        sys.path.append(path)
    else : pass

# VIF 산출을 위해 Preprocessing 모듈의 calculate_vif 메소드 사용 
# Label Encoding을 위해 Preprocessing 모듈의 Encoder 클래스 사용

syspath('./dataset/mod')

import Preprocessing
from Preprocessing import Encoder

# 데이터셋 불러오기
df = pd.read_pickle('./dataset/20_basic_feature_df.pkl')

# 분석을 위하여 Label Encoding 진행
en = Encoder(df)
le = en.label()
le.head()

# Label Encoding 한 데이터에서 LABEL 칼럼 제거
X1 = le.drop('LABEL', axis=1)

# 모듈을 사용하여 VIF 산출
Preprocessing.calculate_vif(X1)

# VIF가 가장 높은 컬럼 제거 : GDS_GROSS_MASS_35	(3402.018726)

X2 = X1.drop('GDS_GROSS_MASS_35', axis=1)

Preprocessing.calculate_vif(X2)

# VIF가 가장 높은 컬럼 제거 : COR_FINANCIAL_VALUE	(243.665473)
# 현재 GDS_GROSS_MASS_35 / COR_FINANCIAL_VALUE 제거 됨

X3 = X2.drop('COR_FINANCIAL_VALUE', axis=1)

Preprocessing.calculate_vif(X3)

# VIF가 가장 높은 컬럼 제거 : IMP_TRANSPORT_MODE_AT_BODR_26	(116.447418)
# 현재 GDS_GROSS_MASS_35 / COR_FINANCIAL_VALUE / IMP_TRANSPORT_MODE_AT_BODR_26 제거 됨

X4 = X3.drop('IMP_TRANSPORT_MODE_AT_BODR_26', axis=1)

Preprocessing.calculate_vif(X4)

# VIF가 가장 높은 컬럼 제거 : TRD_COUNTRY_2	(22.609580)
# 현재 GDS_GROSS_MASS_35 / COR_FINANCIAL_VALUE / IMP_TRANSPORT_MODE_AT_BODR_26 / TRD_COUNTRY_2 제거 됨

X5 = X4.drop('TRD_COUNTRY_2', axis=1)

Preprocessing.calculate_vif(X5)

# 모든 칼럼들이 VIF 20 이하가 된 것을 확인 가능, VIF 산출 종료
