==> 22_Under_Sampling.ipynb <==
# Under Sampling

- Under Samping에서 sample은 다수 클래스의 모집단 특성을 반영해야 한다

- 그러려면 충분한 수의 sample (sampling된 sample의 최대 수는 소수 클래스의 sample 수)이 있어야 함

- 하지만 우리 팀 데이터는 명목 변수에 값이 너무 많아 소수 클래스의 sample수로는 값을 모두 커버 불가능

- 따라서 현 상태로는 Under sampling이 불가능하다고 판단됨

    - 단, 본 툴에 데이터셋을 넣어 sampling이 모집단 분포를 충분히 반영할 수 있으리라 판단되는 수준에 이른다고 보이면 Under Sampling 사용 가능
    

def syspath(path) :
  if path not in sys.path :
    sys.path.append(path)
  else : pass

import numpy as np
import pandas as pd
import warnings
import sys
warnings.filterwarnings('ignore')

syspath('./dataset/mod')

import Preprocessing

# 소수 클래스 (LABEL = 1)의 수 보다 feature의 unique 값들이 훨씬 큼
# 따라서 Under Sampling은 사용하기 힘들다고 판단
# 모듈 EDA의 usct 메서드 사용 (메서드 내용은 모듈 확인)

Preprocessing.usct('./dataset/20_basic_feature_df.pkl')
