==> 23_Over_sampling.ipynb <==
from imblearn.over_sampling import SMOTE
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
import warnings
import sys
warnings.filterwarnings('ignore')

def syspath(path) :
    if path not in sys.path :
        sys.path.append(path)
    else : pass

syspath('./dataset/mod')

from Preprocessing import Encoder

# 데이터 로드 후 Label Encoding 실시
odf = pd.read_pickle('./dataset/00_feature_with_label.pkl')
en = Encoder(odf)
df = en.label()

from sklearn.model_selection import GridSearchCV
from lightgbm import LGBMClassifier
from sklearn.metrics import accuracy_score, precision_score, recall_score,\
confusion_matrix, f1_score, roc_auc_score

def get_clf_eval(y_test, pred, pred_proba):
    confusion = confusion_matrix(y_test, pred)
    accuracy = accuracy_score(y_test, pred)
    precision = precision_score(y_test, pred)
    recall = recall_score(y_test, pred)
    f1 = f1_score(y_test, pred)
    # ROC AUC는 pred 대신 pred_proba
    roc_score=roc_auc_score(y_test,pred_proba) # pred_proba

    print('오차행렬')
    print(confusion)
    print('정확도:{0:.4f}, 정밀도:{1:.4f}, 재현율:{2:.4f},\
    f1 score:{3:.4f}, ROC AUC:{4:.4f}'.format(accuracy, precision, recall, f1, roc_score))
    

def do_lgbm(X_train, X_test, y_train, y_test, param=1, param_compare=0):
    if param == 1:
        lgbm_clf = LGBMClassifier(n_estimators=100, random_state=156, n_jobs=-1)
    elif param == 2:
        lgbm_clf = LGBMClassifier(n_estimators=100, random_state=156, num_leaves=128,
                              n_jobs=-1)
    elif param == 3:
        lgbm_clf = LGBMClassifier(n_estimators=100, random_state=156, num_leaves=128,
                              max_depth=64, n_jobs=-1)
    elif param == 4:
        lgbm_clf = LGBMClassifier(n_estimators=100, random_state=156,
                              max_depth=32, min_child_samples=64,
                              num_leaves=96, n_jobs=-1)

    evals = [(X_test, y_test)]
    
    if param_compare==0:
        lgbm_clf.fit(X_train,y_train, early_stopping_rounds=10,eval_set=evals, \
                      eval_metric='logloss',verbose=True)
    elif param_compare==1:
        lgbm_clf.fit(X_test, y_test, early_stopping_rounds=10,eval_set=evals, \
                      eval_metric='logloss',verbose=True)
    w_preds = lgbm_clf.predict(X_test)
    w_preds_proba = lgbm_clf.predict_proba(X_test)[:,1]

    get_clf_eval(y_test, w_preds, w_preds_proba)

 #### 아래와 같이 smote 수행 시 정밀도가 큰 폭으로 떨어지는 것을 볼 수 있음. 이유는 하단에서 추가 설명

    오차행렬
    [[422328   4213]
    [   339   2422]]
    정확도:0.9894, 정밀도:0.3650, 재현율:0.8772,    f1 score:0.5155, ROC AUC:0.9948

ndf = df.copy()
X_feature = ndf.drop('LABEL', axis=1)
y_label = ndf.LABEL

X_train, X_test, y_train, y_test = train_test_split(X_feature,
                                                    y_label, test_size = 0.2, random_state=50)

sm = SMOTE(random_state=50, n_jobs=-1)

X_resampled1, y_resampled1 = sm.fit_resample(X_train,list(y_train))
do_lgbm(X_resampled1, X_test, y_resampled1, y_test, 1)

# svmsmote는 실행 시간이 오래걸려(최소 6시간) 별도 pkl 파일 중 일부를 추출해 놓음
import joblib
df_svm = joblib.load('/content/drive/MyDrive/NA1/Project2_Final/dataset/SVMSMOTE1.pkl')
df_svm

## svmsmote 및 smote 등 오버샘플링을 사용하지 않는 이유

- 하단 2개의 결과 셀에서 볼수 있는 것처럼, 원본 데이터프레임 중 TRD_COUNTRY_2의 유니크 값의 범위가 0~ 108임
- 반면에 svmsmote가 적용된 새로운 데이터프레임 중 TRD_COUNTRY_2의 유니크 값의 범위는 -18 ~ 130임
=> 즉 명목형 변수임에도 새로운 값이 창조됨

이로인해 오히려 정밀도가 하락하는 현상이 발생했고, 이러한 이유로 본 프로젝트에서는 오버샘플링을 사용하지 않음.

# 원본 데이터프레임 중 TRD_COUNTRY_2의 unique 리스트
origin_unique_list = df.TRD_COUNTRY_2.unique()
origin_unique_list.sort()
origin_unique_list

# svmsmote를 수행한 새로운 데이터프레임 중 TRD_COUNTRY_2의 unique 리스트
svmsmote_unique_list = df_svm.TRD_COUNTRY_2.unique()
svmsmote_unique_list.sort()
svmsmote_unique_list
